backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
aniso8601==9.0.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
blobfile==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
distlib==0.3.6
django-cors-headers==3.13.0
dnspython==2.3.0
filelock==3.9.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fonttools==4.38.0
frozenlist==1.3.3
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.65
loguru==0.6.0
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.6.3
multidict==6.0.4
numpy==1.24.1
openai==0.27.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pinecone-client==2.1.0
platformdirs==2.6.2
plotly==5.12.0
pycryptodomex==3.16.0
pydantic==1.10.4
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.0
scipy==1.10.0
six==1.16.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
sqlparse==0.4.3
tenacity==8.1.0
threadpoolctl==3.1.0
tiktoken==0.3.1
tqdm==4.64.1
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
virtualenv==20.17.1
Werkzeug==2.2.2
win32-setctime==1.1.0
yarl==1.8.2
